requirements.txt
basedosdados
ckanapi
google-cloud-bigquery-storage
Jinja2
pandas-gbq
pyaml
pyarrow
ruamel.yaml
tomlkit
tqdm
click
google-cloud-bigquery
google-cloud-storage
pandas
streamlit
pandas
numpy
pillow
basedosdados==0.1.0
google-cloud-bigquery==1.28.0
streamlit
pandas
numpy